add bottom nav styling
