+++ client/src/App.less
@@ -1,4 +1,5 @@
 @import "~antd/dist/antd.less";
+@import "~antd-mobile/dist/antd-mobile.less";
 
 .App {
   text-align: center;
