+++ client/src/components/Navigation/styled.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const NavContainer = styled.div`
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+`;
+
+export default NavContainer;
